Manejo de stock fraccionable

+++ database/migrations/2025_01_27_154900_create_articulos_table.php
@@ -42,9 +42,9 @@ return new class extends Migration {
             $table->text('descripcion')->nullable();
             $table->decimal('costo');
             $table->decimal('precio')->nullable();
-            $table->unsignedInteger('stock')->default(0);
-            $table->unsignedInteger('abiertos')->default(0);
-            $table->unsignedInteger('mermas')->default(0);
+            $table->decimal('stock')->default(0);
+            $table->decimal('abiertos')->default(0);
+            $table->decimal('mermas')->default(0);
             $table->boolean('fraccionable')->default(false);
             $table->timestamps();
             $table->softDeletes();
